pyproject.toml
[project]
name = "social-media-agency"
version = "0.1.0"
description = "Internal social media agency MVP — CrewAI content engine + Streamlit dashboard + SQLite"
requires-python = ">=3.12,<3.14"
dependencies = [
    "crewai[google-genai]>=0.108.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.40.0",
    "pandas>=2.2.0",
    "pillow>=11.3.0",
    "plotly>=6.6.0",
    "reportlab>=4.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
